Switch teams with realm actions
